7143d879002631f9d406efa3a1b7ff5b80679422 merkle: constify a bunch of functions (Andrew Poelstra)
132e77d60934a8f92ecbe914358e3f4d146ef2c8 type_name: add tmr() function, equality comparison with Final (Andrew Poelstra)
ed585794b989abadd486e86ccbd99e71762beb7b value: eliminate some casts from Word (Andrew Poelstra)
82cf682ad8caef7828fd776c7a9ec0325499e08d value: add Ctx8 costructor from a sha256 hash engine (Andrew Poelstra)

Pull request description:

  Ahead of adding "quoting" functionality where we produce Simplicity fragments that compute the CMRs of other Simplicity fragments, constify a few functions, cleanup some others, and add `Value` constructors from SHA256 contexts.
  
  The next PR will use the new bitcoin-hashes release from https://github.com/rust-bitcoin/rust-bitcoin/pull/6646 which will let us do some more cleanups and constification, and then we can start doing "real" changes.


ACKs for top commit:
  stringhandler:
    ACK 7143d879002631f9d406efa3a1b7ff5b80679422 ran cargo test


Tree-SHA512: 0f323c9e4b601eb19dc96853727f234079e36120e1f06ca91871dffd94c6e2b044f65cad49553d0969b75f3114fc6816c4c2bfc5f2691a32ba0044d6b0da50e4

+++ src/types/arrow.rs
@@ -314,7 +314,7 @@ impl<'brand> CoreConstructible<'brand> for Arrow<'brand> {
     fn const_word(inference_context: &Context<'brand>, word: Word) -> Self {
         Arrow {
             source: Type::unit(inference_context),
-            target: Type::two_two_n(inference_context, word.n() as usize), // cast safety: 32-bit machine or higher
+            target: Type::two_two_n(inference_context, word.n()),
             inference_context: inference_context.shallow_clone(),
         }
     }
